fix(settings): removed data-testid, moved side panel titles, descriptive test titles
