+++ src/app/devices/views/DeviceList/DeviceList.tsx
@@ -8,14 +8,13 @@ import DeviceListTable from "./DeviceListTable";
 
 import PageContent from "@/app/base/components/PageContent";
 import { useFetchActions, useWindowTitle } from "@/app/base/hooks";
-import { useSidePanel } from "@/app/base/side-panel-context";
+import { getSidePanelTitle, useSidePanel } from "@/app/base/side-panel-context";
 import DeviceHeaderForms from "@/app/devices/components/DeviceHeaderForms";
 import { deviceActions } from "@/app/store/device";
 import deviceSelectors from "@/app/store/device/selectors";
 import { FilterDevices } from "@/app/store/device/utils";
 import type { RootState } from "@/app/store/root/types";
 import { tagActions } from "@/app/store/tag";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 
 const DeviceList = (): React.ReactElement => {
   const dispatch = useDispatch();
