+++ src/app/domains/views/DomainsList/DomainsList.tsx
@@ -8,10 +8,9 @@ import DomainsTable from "./DomainsTable";
 import { useZones } from "@/app/api/query/zones";
 import PageContent from "@/app/base/components/PageContent";
 import { useWindowTitle } from "@/app/base/hooks";
-import { useSidePanel } from "@/app/base/side-panel-context";
+import { getSidePanelTitle, useSidePanel } from "@/app/base/side-panel-context";
 import DomainForm from "@/app/domains/components/DomainForm";
 import domainsSelectors from "@/app/store/domain/selectors";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 
 const DomainsList = (): React.ReactElement => {
   const domains = useSelector(domainsSelectors.all);
