+++ src/app/images/components/ImagesForms/SelectUpstreamImagesForm/SelectUpstreamImagesForm.tsx
@@ -199,10 +199,7 @@ const SelectUpstreamImagesForm = (): ReactElement => {
   };
 
   return (
-    <div
-      className="select-upstream-images-form"
-      data-testid="select-upstream-images-form"
-    >
+    <div className="select-upstream-images-form">
       Select images to be imported and kept in sync daily. Images will be
       available for deployment on MAAS managed machines.
       {tooManySources && (
