+++ src/app/kvm/utils.ts
@@ -4,8 +4,8 @@ import { KVMSidePanelViews } from "./constants";
 import type { KVMStoragePoolResource, KVMStoragePoolResources } from "./types";
 
 import type { SidePanelContent } from "@/app/base/side-panel-context";
+import { getSidePanelTitle } from "@/app/base/side-panel-context";
 import type { Pod } from "@/app/store/pod/types";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 
 /**
  * Returns a string with the formatted byte value and unit, e.g 1024 => "1KiB"
