+++ src/app/pools/views/PoolsList.tsx
@@ -2,14 +2,13 @@ import type { ReactElement } from "react";
 import { useEffect } from "react";
 
 import PageContent from "@/app/base/components/PageContent";
-import { useSidePanel } from "@/app/base/side-panel-context";
+import { getSidePanelTitle, useSidePanel } from "@/app/base/side-panel-context";
 import AddPool from "@/app/pools/components/AddPool/AddPool";
 import DeletePool from "@/app/pools/components/DeletePool/DeletePool";
 import EditPool from "@/app/pools/components/EditPool/EditPool";
 import PoolsListHeader from "@/app/pools/components/PoolsListHeader";
 import PoolsTable from "@/app/pools/components/PoolsTable";
 import { PoolActionSidePanelViews } from "@/app/pools/constants";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 import { isId } from "@/app/utils";
 
 const PoolsList = (): ReactElement => {
