+++ src/app/preferences/views/SSHKeys/views/SSHKeysList.tsx
@@ -3,12 +3,11 @@ import { useEffect } from "react";
 
 import PageContent from "@/app/base/components/PageContent";
 import { useWindowTitle } from "@/app/base/hooks";
-import { useSidePanel } from "@/app/base/side-panel-context";
+import { getSidePanelTitle, useSidePanel } from "@/app/base/side-panel-context";
 import AddSSHKey from "@/app/preferences/views/SSHKeys/components/AddSSHKey/AddSSHKey";
 import DeleteSSHKey from "@/app/preferences/views/SSHKeys/components/DeleteSSHKey";
 import SSHKeysTable from "@/app/preferences/views/SSHKeys/components/SSHKeysTable/SSHKeysTable";
 import { SSHKeyActionSidePanelViews } from "@/app/preferences/views/SSHKeys/constants";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 import { isId } from "@/app/utils";
 
 type SSHKeysListProps = {
