+++ src/app/subnets/views/FabricDetails/FabricDetails.tsx
@@ -12,13 +12,12 @@ import ModelNotFound from "@/app/base/components/ModelNotFound";
 import PageContent from "@/app/base/components/PageContent";
 import SectionHeader from "@/app/base/components/SectionHeader";
 import { useGetURLId, useWindowTitle } from "@/app/base/hooks";
-import { useSidePanel } from "@/app/base/side-panel-context";
+import { getSidePanelTitle, useSidePanel } from "@/app/base/side-panel-context";
 import { fabricActions } from "@/app/store/fabric";
 import fabricSelectors from "@/app/store/fabric/selectors";
 import { FabricMeta } from "@/app/store/fabric/types";
 import type { RootState } from "@/app/store/root/types";
 import { subnetActions } from "@/app/store/subnet";
-import { getSidePanelTitle } from "@/app/store/utils/node/base";
 import subnetURLs from "@/app/subnets/urls";
 import { isId } from "@/app/utils";
